+++ server/server_app/src/main/java/com/example/jphacks_server/service/LoginService.java
@@ -1,0 +1,106 @@
+package com.example.jphacks_server.service;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.example.jphacks_server.entity.Users;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Service;
+
+import java.util.Date;
+import java.util.List;
+
+@Service
+public class LoginService {
+
+    HttpHeaders responseHeaders = new HttpHeaders();
+
+    public LoginService(){
+        responseHeaders.setContentType(MediaType.APPLICATION_JSON);
+    }
+
+
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    private static final Long EXPIRATION_TIME = 1000L * 60L * 10L;
+
+    public ResponseEntity<String> login(Users usersData){
+        String query = "SELECT * from users where users_login_id = ? and users_login_password = ?";
+        List<Users> users = jdbcTemplate.query(query,new BeanPropertyRowMapper<>(Users.class), usersData.getUsersLoginId(), usersData.getUsersLoginPassword());
+        String token = null;
+        System.out.println(users.size());
+        if(users.size() != 0){
+            String usersId = String.valueOf(users.get(0).getUsersId());
+            token = createToken(usersId);
+        }
+
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode root = mapper.createObjectNode();
+
+        if(token == null){
+            root.put("status", "failed");
+            responseHeaders.remove("Authorization");
+        }else{
+            root.put("status", "success");
+            responseHeaders.set("Authorization", token);
+        }
+
+        ResponseEntity<String> responseEntity = null;
+        try {
+            responseEntity = new ResponseEntity<String>(mapper.writeValueAsString(root), responseHeaders, HttpStatus.OK);
+        } catch (
+                JsonProcessingException e) {
+            e.printStackTrace();
+        }
+
+        return  responseEntity;
+
+    }
+
+
+
+    private String createToken(String subject) {
+
+
+        String secretKey = "panda";
+        Date issuedAt = new Date();
+        Date notBefore = new Date(issuedAt.getTime());
+        Date expiresAt = new Date(issuedAt.getTime() + EXPIRATION_TIME);
+
+
+        Algorithm algorithm = Algorithm.HMAC256(secretKey);
+
+        try {
+            String token = JWT.create()
+                    // registered claims
+                    //.withJWTId("jwtId")        //"jti" : JWT ID
+                    //.withAudience("audience")  //"aud" : Audience
+                    //.withIssuer("issuer")      //"iss" : Issuer
+                    .withSubject(subject)         //"sub" : Subject
+                    .withIssuedAt(issuedAt)      //"iat" : Issued At
+                    .withNotBefore(notBefore)    //"nbf" : Not Before
+                    .withExpiresAt(expiresAt)    //"exp" : Expiration Time
+                    //private claims
+                    //private claims
+                    //.withClaim("X-AUTHORITIES", "aaa")
+                    //.withClaim("X-USERNAME", "bbb")
+                    .sign(algorithm);
+            System.out.println("generate token : " + token);
+            return token;
+        } catch (Exception e){
+            e.printStackTrace();
+        }
+
+        return null;
+    }
+}
