+++ server/server_app/src/main/java/com/example/jphacks_server/service/UsersService.java
@@ -16,9 +16,9 @@ import java.util.List;
 public class UsersService {
 
     @Autowired
-    private JdbcTemplate jdbcTemplate;
+    JdbcTemplate jdbcTemplate;
 
-    private static final Long EXPIRATION_TIME = 1000L * 60L * 10L;
+
 
     public List<Users> findAll() {
         // 実行する SQL を組み立てて実行
@@ -27,51 +27,6 @@ public class UsersService {
         return users;
     }
 
-    public String login(String id, String password){
-        String query = "SELECT * from users where users_login_id = ? and users_login_password = ?";
-        List<Users> users = jdbcTemplate.query(query,new BeanPropertyRowMapper<>(Users.class), id, password);
-        String token = null;
-        System.out.println(users.size());
-        if(users.size() != 0){
-            String usersId = String.valueOf(users.get(0).getUsersId());
-            token = createToken(usersId);
-        }
-        return token;
-    }
 
 
-    private String createToken(String subject) {
-
-
-        String secretKey = "panda";
-        Date issuedAt = new Date();
-        Date notBefore = new Date(issuedAt.getTime());
-        Date expiresAt = new Date(issuedAt.getTime() + EXPIRATION_TIME);
-
-
-        Algorithm algorithm = Algorithm.HMAC256(secretKey);
-
-        try {
-            String token = JWT.create()
-                    // registered claims
-                    //.withJWTId("jwtId")        //"jti" : JWT ID
-                    //.withAudience("audience")  //"aud" : Audience
-                    //.withIssuer("issuer")      //"iss" : Issuer
-                    .withSubject(subject)         //"sub" : Subject
-                    .withIssuedAt(issuedAt)      //"iat" : Issued At
-                    .withNotBefore(notBefore)    //"nbf" : Not Before
-                    .withExpiresAt(expiresAt)    //"exp" : Expiration Time
-                    //private claims
-                    //private claims
-                    //.withClaim("X-AUTHORITIES", "aaa")
-                    //.withClaim("X-USERNAME", "bbb")
-                    .sign(algorithm);
-            System.out.println("generate token : " + token);
-            return token;
-        } catch (Exception e){
-            e.printStackTrace();
-        }
-
-        return null;
-    }
 }
